Submitted by: George Bosilca Reviewed by: Jeff Squyres
Patches to make SM PTL work properly: use the same match functionality
as other PTLs and some memory fixes.  This also fixes Probe/Iprobe
functionality for the sm PTL.

Rich has voiced concern about the performance of this, but I'm making
an executive decision to get correct functionality so that we can
release the beta.  George has tested and assures me that it actually
gets slightly better performance than the old method, but we can work
out these details post-beta.

(note: these same issues are on the trunk; this patch will fix them,
but has not been applied because of the potential performance
implications / objections.  To be discussed on the mailing list...)