Submitted by: Ralph Castain Reviewed by: Jeff Squyres
Yay svn merge -- brought this directly from the trunk because it's two
patches and involved property changes:

svn merge -r 5979:5980 https://svn.open-mpi.org/svn/ompi/trunk
svn merge -r 5982:5983 https://svn.open-mpi.org/svn/ompi/trunk

Commit message from 5980:
=========================

Fix the registry search routine to remove a buffer that wasn't
expanding as it should - cause of recent problems observed when
spawning larger numbers of processes.

For anyone interested, the problem stemmed from two things:

1. a bug in the ompi_bitmap utility (which I copied to orte_bitmap to
avoid unintentionally disturbing something else) that causes the
bitmap NOT to expand unless the caller asks for a bit that is more
than one byte outside the current array size. The unit test didn't
pick it up because it doesn't check that close to the boundary.

2. a "feature" in the ompi_bitmap utility that only expands the array
if you try to SET a bit outside the current boundary, but NOT if you
try to CLEAR a bit outside the array limit. This appears intentional
as the unit test checks for this behavior, but I hadn't been expecting
the asymmetry.

The orte_bitmap utility now appropriately expands in both
circumstances. I also added a function to expand the array so it
"covers" a bit location without setting or clearing it. The function
allows you to ensure the array is big enough to handle the specified
bit, but leave the bit alone if it already is there (the other
functions would set/clear it if it was).

I've tested it with up to 100 processes without problem.


Commit message from 5983:
=========================

Fix a couple of minor thiings per Jeff's suggestions:

1. Change the hardcoded "8" to "SIZE_OF_CHAR"

2. Change "cover_bit" to "resize"

Also removed an extra "if" check that was duplicative.

Again, tested on up to 100 processes.