Submitted by: Josh Hursey Reviewed by: Jeff Squyres
Fix for a problem Edgar and Jeff identified WRT PLS determining if we
are oversubscribed on a node. And thus whether to call sched_yield or
not.

The value of node->node_slots_inuse does not currently represent the
number of slots actually in use, at the moment. This is actually a bug
in the RAS/RMAPS base components, but the fix for that specific bug is
bigger than we want to address at the moment (but will certianly do so
in the near future).

Since we cannot trust this value, use the total number of mapped
processes (which was properly set by the RMAPS component upon mapping
-- Just not properly propagated back to the registry's node segment)
from the process mapping.

In addition to this change I cleaned up a couple of the debug
messages. It seems that TM and RSH are the only two directly effected
by this. SLURM would be if that section of code wasn't currently
inactive, but put the fix in for prosparity.