Submitted by: George Bosilca Reviewed by: Jeff Squyres (sort of)
This is a rollup of ddt commits from the trunk:

r8004:
Do not add epsilon to the data extent if there is a user set UB for
the data.

r8005:
Make a compiler happy about casting.

r8006:
This is the way to compute the position for a convertor under the new
rules.  This file is now yet activated. It will became the default
after the next commit.  (checkpoint to start testing on other
clusters)

r8019:
When we advance the convertor by a multiple of the data size there is
a quick optimization.  We can compute the number of complete datatype
that we will advance, update the stack and then compute the new
position taking in acount only the remaining bytes.

r8020:
Several bug fixes:
- if the alignment of wchar is zero then wchar_t is not supported by
  the OS. We skip it.
- Now that the definition of end_loop change compute the
  first_elem_description for all predefined datatypes.
- In debug mode print a list of the datatypes that are not supported
  by the current architecture.

r8021:
The shiny new pack and unpack functions. The big difference is that
the displacement is never stored on the stack. It is partially stored
on the stack depending on the loops but every time we pack/unpack a
basic datatype we take in account again it's displacement.  This
approach make the whole logic a lot simpler. In same time I split the
big functions in several basic block.

r8022:
Force the data name to the empty string when we call destroy.

r8023:
Add some macros in the begining of the file. They are not used right
now, but they will be in few days.

Do not ignore the type and extent of the last optimized basic type in
some special cases.

Update the last fake END_LOOpP with the correct value for the
first_elem_disp field.

r8024:
The END_LOOP structure change the name of one of it's fields.

r8025:
Add one prototype from the new_position.c

r8026:
Slim fast ... Do as less as possible on the critical path. The most
expensive function now is the one that compute the stack when we move
to a new position. For this function there are several versions
depending on the type of the data annd the position where we want to
go.

Change the name of one of the fields in the end_loop structure.
Update all the macros to reflect the change.
A slightly different version of the boundaries checking function.

r8028:
Take in account the original displacement for contiguous datatypes.
Limit the amount of data to be packed to the remaining on the
convertor. This make the things a lot simpler in the pack/unpack
functions.

r8029:
And here is the makefile that integrate the new files. Now ... have as
much fun as I did :)

r8030:
Shame on me ... there should be extent not displacement.

r8031:
A step forward. The original displacement for contiguous data with
gaps is now correctly computed. At least the original displacement.

r8034:
Change some variable names (from pSrc to something more clear like
user_memory and/or packed_buffer).

r8040:
When we finish one of the loops take care of the index of the begining
of the loop. If it's -1 then we just complete the full datatype
... therefore we have to do something special.

r8041:
Change some variable's names.

r8042:
A simpler way to get output about the packing/unpacking. Now there are
2 MCA parameters datatype_pack_debug and datatype_unpack_debug. When
they are set to 1 the ddt engine will dump a lot of messages. Dont
turn them to one by default. But if you notice any problems in the ddt
you can turn them to one and send me the output.

First step toward adding memory to the convertor. It will be able to
keep partial basic datatype between calls ...

r8047:
By default there is no pending length on the convertor.

r8048:
Allow the convertor to remember the last trucated unpack. If the same
convertor is used for the next unpack it will put the data back
correctly. However, if the BTL/PTL create a new convertor, even if it
clone the last one this magic will not happens !

r8051:
More optimization of the data-type description are now possibles. Some
corner cases are corrected. As a result we discover more accurately
the contiguous part of the data memory layout.

r8054:
Do not do any local copy into the storage if this convertor is
finished. This is usefull in the Pack/Unpack case when there is more
data in the packed buffer than the one we try to extract.

r8056:
Turn off the missing data-type message on MPI_Init.

r8069:
Activate the protection against unavailable datatypes. They get a flag
DT_FLAG_UNAVAILABLE. We check now this flag in all the send/recv
operations via the macros on mpi/c/bindings.h.

This flag is inherited by all datatypes create with unavailable
datatypes. Basically, we let the user create the wrong datatype but we
dont let him using it for any pt2pt communications or any pack/unpack.

r8080:
Somehow I miss to remove one of the previous definition for the
unavailable data.