Submitted by: Jeff Squyres Reviewed by: Brian Barrett
Cumulative merge from the trunk of several r numbers to bring
correctness back to the sm btl:

r8134:
- Add copyright headers to btl_sm_frag.h
- Ensure to convert base_shared_mem_flags to be a relative offset in
  the global storage, and then to convert that back to an absolute
  virtual address before we try to use it
- Don't double increment n_local_procs when calculating the peer rank
  during bootstrapping of the different base address case

Something else is still wrong; if mmap() returns a different base
address, things don't work (i.e., segv or hang forever when you try to
send a message).  More specifically, the bootstrapping now seems to
correctly handle the case when mmap() base addresses are different,
but the message passing does *not* -- it always assumes that the
mmap() base addresses are the same.

Still working on the fix for that -- want to checkpoint what has been
done so far to facilitate working on different machines...

r8135:
Next checkpoint in the sm btl fixes:

- Add big comment about a general overview of what the sm btl is doing
- random small code cleanups
- fix instances of mca_btl_sm[0] to mca_btl_sm[1] where relevant
- remove a lot of unused, confusing, and incorrect interface functions
  from ompi_fifo.h and ompi_circular_buffer.h.  These functions, if
  they were used, would not work properly with the scheme that the sm
  btl uses with the fifos (i.e., receiver makes right -- if necessary)
- add some missing offset computations in the fifo and circular
- buffers
- change the types of offsets to be ssize_t, not size_t
- remove an offset parameter from a function that didn't need it

r8136:
More changes for correctness of the sm btl.
- remove dead code that isn't used anywhere (originally ompi_fifo_t
  was going to be a generalized class, but now it's exclusively used
  in the sm stuff, so there's no point in the generalized code that
  definitely *won't* work with the sm btl, or is not being used now
  [SVN always has history so we can go back])
- had to add an interprocess lock in the area where the writer may
  create a new circular buffer to ensure that the reader's tail
  doesn't accidentally end up back in the same old buffer while the
  head continues on to a new circular buffer (this was what was
  happening to cause some intel tests to hang -- e.g., MPI_Scan_c,
  MPI_Send_fairness_c and MPI_Isend_fairness_c).  Unbelievably, this
  may actually *increase* performance because it may order things
  better.  Will do performance testing tomorrow.  We're fairly certain
  that this lock can probably be removed and the code fixed in a
  different way, but we're under a deadline and correctness comes
  first, so it's been added to the to-do list to come back and
  re-examine this case later.

r8137:
Don't run ompi_fifo and ompi_circular_buffer tests; the interfaces
have changed and the tests have not changed with them.

r8138:
Add documentation about what is happening in this class.

r8139:
Remove debugging printf