Submitted by: George Bosilca Reviewed by: Jeff Squyres
Add a MCA params to allow/disable the MX shared memory
capabilities. Right now this param is labeled as internal so the users
will not see it but it is not read-only so we can still play with it
(that's for our internal tests). This is supposed to dissapear later
after the next (or next next) release of the MX library, but we need
it now as a quick fix before the release.