Submitted by: Edgar Gabriel Reviewed by: Jeff Squyres
Merge r8243, 8244, and 8245 from the trunk to the v1.0 branch.

r8243:
fixing the fortran interface for MPI_Get_address. Since the address
argument is defined to be 

integer (KIND=MPI_ADDRESS_KIND)

we have to map that to an MPI_Aint in the f->c interface, else the
typecast back from the C to the Fortran interface will be wrong as
well.  The same holds (at least) for all other new MPI-2 datatype
functions which take a byte-displacement or an address as an argument.

r8244:
fixes for other routines which take an MPI-2 byte displacement or an
address as an argument.

r8245:
Remove a little extra memory that is no longer needed