Fixed error when count = 0 was given for multiple completion MPI functions (MPI_TESTSOME, MPI_TESTANY, MPI_TESTALL, MPI_WAITSOME, MPI_WAITANY, MPI_WAITALL).
r9504:
Accept an array of request set to NULL if the count is set to ZERO.

r9512:
Follow the standard when the count is set to zero. Or let me say it differently,
follow the statement from the standard that make the most sense to me. In this
particular case this statement is:
If there is no active handle in the list it returns outcount = MPI_UNDEFINED.

Submitted by: George Bosilca
Reviewed by: Brian Barrett

Fix error in testsome if MPI_STATUSES_IGNORE is passed

Submitted by: Brian Barrett
Reviewed by: George Bosilca