Submitted by: Brian Barrett Reviewed by: Jeff Squyres
Fix for bug #17. Solaris provides stubs (who knows why, but they do)
for the pthreads library that don't do anything but are there when no
special options are given. Both the GNU compiler and the Sun compiler
could sometimes ignore the -K* options, causing badness when building
with posix threads. Don't try those options ;).

We still try -pthread and -pthreads because the compilers *do* error
when they see those options and some versions of the GNU compiler do
understand those flags (and do all the right things in that case).