Submitted by: Galen Shipman Reviewed by: Brian Barrett
Fixes tickets #53 and #28.

r10072
Allow maximum send size to be less than the eager limit. 
Instead of figuring out which free list the fragment belongs to based on size
we simply store a pointer to the list which it belongs in the fragment.

r10078
Track fragments by list, not by size.. 

r10142
track which list a fragment came from, allows returning based on list, not
on size. 

r10157
Fix to allow eager limit and max send size to be any size (within
resource limitations). Instead of storing the ompi_free_list_t * in
the fragment, we use the frag type enum, this tells us where the frag
came from and where it should return.. This could also be done in
mvapi but is not a high priority moving forward..