Submitted by: Jeff Squyres Reviewed by: Brain Barrett, George Bosilca
r10465
Move the Fortran file into the nodist headers.

r10464
Fix the fix -- if we're not compiling the profiling layer, then we
cannot include the PMPI_WTIME|WTICK functions in the external and
double precision statements because some compilers complain about
this.  Instead, we need to use the macro that is defined by
configure.ac (MPIF_H_PMPI_W_FUNCS).  This unfortunately means that we
need to generate mpif.h (in addition to mpif-config.h) because the
"external" statement is toxic to F90 compilers.