Submitted by: George Bosilca, Brian Barrett Reviewed by: Jeff Squyres
r10770 
Create a close function for the mpool SM module, in order to allow the
cleanup. The mca_common_sm_mmap file was left over by the SM mpool,
and there was nobody able to unmap and unlink it.

r10869 
As this object was created using the OBJ_NEW it should be destroyed
using OBJ_RELEASE. 

r10886
 * rename the collectives sm bootstrap area to be consistent with other 
   shared memory segments 
 * make sure to properly unlink the collectives sm bootstrap area at 
   shutdown 
 * Add missing / in the path for the mpool shared memory segment 
 * make sure to release the common_mmap structure in the SM btl 
   after unlinking the file during shutdown