* Submitted by: Graham Fagg  * Reviewed by: Sven Stork
Refs #373

Simple fix (for #363). We push segment size to type size. In other
algorithms we switch of segementing altogether. But really the DDT can
probably handle partial types so we could really keep the segsize
constant (for all but reduce ops) and treat it just as byte arrays..
todos: macroize it as we do it 10 different ways, add mca params to
control handling (push up size, no change, switch off segmenting)