* Submitted by: Brian Barrett  * Reviewed by: David Daniel
Fixes #400

Please move the (soon to be) attached diff to the v1.1 branch. This is
similar to #399, and the diff corresponds to r11428, r11514, r11745,
and r11746, adjusted for the component name changes that occurred
after v1.1. These changes fix a number of issues with the onesided
code, particularily bugs #354, #303, and #325.