* Submitted by: Jeff Squyres  * Reviewed by: Sven Stork
Fixes #416.

Fix some compiler warnings introduced in r11619.  I checked with
George: ompi_ddt_type_size() returns a signed int only because of the
MPI spec; it will never return a negative value.  So casting the
return value out of it to a (uint32_t) is safe, and makes the
comparisons be between two unsigned values.