* Submitted by: Brian Barrett  * Reviewed by: Jeff Squyres (hah!)
Fixes #445
Refs #441

MPI_Fence's stupid "I might be in a fence epoch but you can't tell
until I do something else" rule screws me up again. If we're in a
FENCE, but not in ACCESS | EXPOSE, put us in ACCESS|EXPOSE, as we are
now known we now in a real Fence epoch. Yay silly MPI standards