Ticket #467: Optimize the data-type engine for heterogeneous cases
 * Submitted by: George Bosilca
 * Reviewed by: Jeff Squyres
 * Merged by: Tim Mattox

Merged r11945, r11962, r11972, r11986, which closes #467:

r11945

Fine grain selection for heterogeneous environments. The hetero
version of the conversion function are more complex and costly than a
simple memcpy. Therefore, we want to decrease as much as possible the
usage of these functions.  We now check not only th HOMOGENEOUS flag
on the datatype or convertor, but the bits indicating a type is in
use. If a communication transfert a type having the same
representation on both peers we can use the optimized version of the
conversion.  In same time we build a more accurate conversion table
for each master convertor, based on the minimum differences between
the 2 architectures.

r11962

With the lastest changes for heterogeneous environments we have to
compute the hetero flag a little bit different. The hetero flag is now
attached to a convertor if and only if we can use the optimized
conversion functions. It's a little bit broader than before (the 2
architectures has to be identical).

r11972

Mea culpa ... that was an weird error. When playing with 64 bits types
we have to the constant before using the << operator ... otherwise
everything is 32 bits ...  which does not give the right answer.

r11986

Cast to the uint64_t before playing with the << operator in order to
make sure we keep all usefull informations.