* Submitted by: Ralph Castain  * Reviewed by: Jeff Squyres
Fixes #490, refs #245.

r11140

When a "fork" launch cannot find the executable, the system used to
just return an error. This meant that the state of that process was
never updated in the registry, leaving the counters at the incorrect
levels. As a result, the triggers would never fire to indicate that
the job had been aborted. This left orterun and other orteds/processes
hanging. 

This fix should fix the problem. I will test it on a broader range of
systems forsooth... 

r11141

Finalize the session directory and runtime when the orted exits due to
a failed launch. 

Additional patch:

Remove a duplicate free() in the fork pls.