Closes #693: fixes SEGV after heterogeneous MPI_Comm_spawn
Submitted by brbarret, Reviewed by Galen, Merged by Tim Mattox

Applied patch which corresponds to r12798, r12822, and r12830.

r12798:
Clean up the way procs are added to the global process list after MPI_INIT: 

 * Do not add new procs to the global list during modex callback or 
   when sharing orte names during accept/connect. For modex, we 
   cache the modex info for later, in case that proc ever does get 
   added to the global proc list. For accept/connect orte name 
   exchange between the roots, we only need the orte name, so no 
   need to add a proc structure anyway. The procs will be added 
   to the global process list during the proc exchange later in 
   the wireup process 
 * Rename proc_get_namebuf and proc_get_proclist to proc_pack 
   and proc_unpack and extend them to include all information 
   needed to build that proc struct on a remote node (which 
   includes ORTE name, architecture, and hostname). Change 
   unpack to call pml_add_procs for the entire list of new 
   procs at once, rather than one at a time. 
 * Remove ompi_proc_find_and_add from the public proc 
   interface and make it a private function. This function 
   would add a half-created proc to the global proc list, so 
   making it harder to call is a good thing. 

This means that there's only two ways to add new procs into the
global proc list at this time: During MPI_INIT via the call to
ompi_proc_init, where my job is added to the list and via
ompi_proc_unpack using a buffer from a packed proc list sent
to us by someone else. Currently, this is enough to implement
MPI semantics. We can extend the interface more if we like,
but that may require HNP communication to get the remote proc
information and I wanted to avoid that if at all possible.

r12822:
Instead of an unknown proc list that requires ownership transfer
of data (which, in turn, requires a complex series of locks to
be held during the transfer), use a modex backing store with
backpointers from the proc to the backing store. The proc
structures no longer own the modex data, which greatly simplifies
locking when an unknown proc suddenly becomes known.

r12830:
Corrections for when threading is enabled.