* Submitted by: Jeff Squyres  * Reviewed by: Sven Stork
Fixes #704

Because of r12712, we actually need to allocate one more item than the
value of {{{__n}}} holds. This is not a problem in the first case
because sizeof(int) == sizeof(MPI_Flogical), so no alloc is actually
performed (which is most compilers, and why we haven't been bitten by
this yet).  But the second case -- where sizeof(int) !=
sizeof(MPI_Flogical) -- is definitely a problem and needs the "+1" in
the alloc, or Bad Things will happen.