Closes #752: inactive btl latency fix
Submitted by Brian, reviewed by Galen, Merged by Tim Mattox

r12938:
If a BTL isn't needed due to exclusivity ranking, need to call a matching
inuse decrement for the increment that was at the start of the procs loop.
Otherwise, the inuse count can end up higher than it actually is and a btl
can end up in the progress loop when it isn't active to any peer.