Closes #744: BTL OpenIB port wireup fix
Submitted by gshipman, Reviewed by jsquyres, Merged by Tim Mattox

r12978:
Handle different subnets correctly and multiple nic endpoint
negotiation. This is somewhat limited currently for expample,
if you have 3 ports on Node A and 5 ports on Node B then the
peers will use 3 ports to communicate with each other. 
This is on a subnet basis, so for any pair of nodes we take the 
intersection of the available ports within a subnet. 

We use subnets to determine reachability for lazy connection
establishment. So if Node A and Node B each have two HCA's
(on seperate networks) then the subnet's must be distinct,
otherwise we will try to wire up HCA's on seperate networks.

r12993:
endpoint should use a uint64_t for subnet, as everyone else does..
makes bad things happen when packing into a 64 bit buffer... 

Misc cleanup.

r13105:
call it what it is...
we are looking at subnet_id's and we are counting active ports per subnet.
move subnet count out of procs loop, no need to do it there...