Closes #834: Fix a memory corruption issue within the GPR
Submitted by rhc, Reviewed by Tim Prins, Merged by Tim Mattox

r13300:
Fix a memory corruption problem deep inside the registry when
subscriptions/triggers are processed. The create_value function
will malloc space for the pointers to keyval objects, but
doesn't actually allocate space for the objects themselves.
When constructing the gpr_notify_data object, we forgot to OBJ_NEW
the keyval objects. Since the create_value function didn't
explicitly NULL those memory locations, it just so happened
that there was a non-NULL address in them....which we dutifully
dumped a keyval into. 

This fix includes two parts: (a) we now initialize the keyval
pointer locations to NULL after the malloc, and (b) we now
OBJ_NEW the keyvals prior to storing info in them. 

BTW, in case anyone reads this and wonders why we don't just
OBJ_NEW the keyvals in create_value, the reason is simply that
some places in the code use static keyvals and simply assign
those addresses into the value object's array. So not everyone
wants to OBJ_NEW keyvals - by not forcing it here in create_value,
we give the user the flexibility to do whatever they want.