Closes #843: CM datatype fix
Submitted by Brian, Reviewed by George, Merged by Tim Mattox

r13397:
When receiving a message, need to account for the fact that the
displacement of the first entry might not be the start of the
user's buffer. This is similar to what ompi_convertor_unpack does.
This is the solution for the test case attached to ticket #690.