Closes #938: fix for race condition in fifo code
Submitted by Gleb & Rich, Reviewed by Rich, RM-Approved by Rich

r13635:
Add error callback to sm BTL. Call it when allocation of the
initial circular buffer fails. If cb is already allocated, but
it is full and allocation of additional cb fails, we spin
waiting for receiver to free space in existing cb.

patch applied corresponding to r13968:
Queue_empty is determined by the reader, and is it's local
view.  However, the writer may continue writing to this queue.
The decision to go on to the next cb_fifo is done in an atomic
region, checking the writer's view. The writer also "changes
it's view" in an atomic region protected by the same lock.