Closes #964: temporary fixes for threaded accept/TCP OOB
Submitted by Jeff, Reviewed by Brian and Galen, RM-Approved by tdd

r14133:
Need to fall back on the event listen mode if the MCA parameter
said use the listen thread, but we're not the HNP. This is better
than not starting up any listen mode, which is what we were
doing before :/

r14135:
A race condition exists on the free list of pending connections
because OPAL_FREE_LIST_WAIT/RETURN will not use locks in a
non-threaded build.  Thus, if non-threaded, conditional use of
locks around the OPAL_FREE_LIST_WAIT/RETURN seems to fix the issue.
Tested at 4K processes and seems to work.