Closes #966: make mpi_leave_pinned work for all datatypes
Submitted by gshipman, Reviewed by bosilca, RM-Approved by tdd

patch applied which corresponds to r14174:
When looking up / registering memory for the leave_pinned protocol
we need to use the address as specified by the convertor, not the
address in the request. Even if the convertor reports that the
memory is contiguous the req_addr may point to something like
MPI_BOTTOM for some datatypes with absolute offsets.