Closes #1001: pathscale/libtool C++ fix
Submitted by Jeff, reviewed by Brian, RM-Approved by tdd

r14454:
Fixes ticket #982. 

Thank God for google-able mailing list archives: 

http://www.mail-archive.com/bug-libtool@gnu.org/msg00899.html 

We ran into this exact bug in Libtool that was causing the C++ 
bindings library to be compiled incorrectly (therefore causing
static initializers to not fire properly when in a shared
library, which is the default installation configuration).
Putting in some libtool patches to fix the problem -- will be
mailing the Libtool crowd shortly to ask for a better fix...