Closes #1094: Optional fortran datatype fixes
Submitted by rusraink, Reviewed by George and Jeff, RM-Approved by tdd

r14818:
The optional Fortran datatypes may not be available 
Do not initialize them, if not. 
If initializing them, check for the correct C-equivalent type 
to copy from... 
Issue a warning, when a type (e.g. REAL*16) is not available to 
build the type (here COMPLEX*32). 
This fixes issues with ompi and pacx. 

Works with intel-compiler and FCFLAGS="-i8 -r8" on ia32.