Closes #1097: Support of message queues for parallel debuggers.
Submitted by paklui, Reviewed by MTT, RM-Approved by tdd

Merged changes from the paklui-1.2-pml tmp branch with this command:

 svn merge -r 15972:16167 https://svn.open-mpi.org/svn/ompi/tmp/paklui-1.2-pml

That tmp branch had changes corresponding to these changesets:

r14789: added suport for full path launching with TotalView

r14790:
Fix a compiler warning - strncmp returns an int, so you have to
compare to 0 instead of NULL.

r15142:
Allow the symbols required by TotalView to be exported even when
the visibility feature is on.

r15346:
This patch bring full support for message queues in Open MPI.
Now the send and receive queues are shared among all PMLs,
they are declared in the base PML, and the selected PML is in
charge of initializing and releasing them. 

The CM PML is slightly different compared with OB1 or DR.
Internally it use 2 different types of requests: light and
heavy. However, now with this patch both types of requests are
stored in the same queue, and cast appropriately on the
allocation macro. This means we might use less memory than we
allocate, but in exchange we got full support for most of the
parallel debuggers. 

Another thing with this patch, is that now for all PML (CM included)
the basic PML requests start with the same fields, and they are
declared in the same order in the request structure. Moreover, the
fields have been moved in such a way that only one volatile/atomic
will exist per line of cache (hopefully).

r15347:
A more consistent version. As we now share the send and receive queue, we
have to construct/destruct only once. Therefore, the construction will
happens before digging for a PML, while the destruction just before
finalizing the component.

r15348:
Looks like MAX is not a standard macro. Anyway, that the heavy
requests is larger than the thin seems to be a "correct" assumption.

r15360: Fix a copy&paste typo.

r15361: Correct the thin and heavy requests management for the CM PML.

r15376:
Don't allow the BTL to be closed before the end of the process. Count the
number of times the BTLs are opened, and then don't remove them until
close was called the same number of times.

r15425:
Fix the problem reported on #1087. The global send and receive
requests queues are now release in the base close, so there is
no need for the cm PML to destroy them.

r15716:
 * Remove the logic for Solaris to always use the FreeBSD version of qsort.
 * Give user the option to configure with the broken qsort fix instead
   of using the native qsort.

r15915:
Don't show the requests having a negative tag (they are internal
requests for Open MPI). Add a variable to allow Open MPI developers
to see all internal messages.

r15970:
Temporary workaround for making Totalview be able to get those opal
symbols and load into the library when compiled with a Sun Studio C compiler

r16027:
complete the fix started in r15915, that is, to prevent negative tags
from showing up in the message queue graph. Tags are now casted to int
before the negative checks, since tags by the spec are stored as
mqs_tword_t, an unsigned long long.

r16051: Nothing relevent (yet). Just making debugging more enjoyable.

r16052:
Fix a problem with OPAL_ALIGN that causes the upper bytes to get chopped
off and bogus addresses to show up for the requests, which in turns causes
message queues not showing up when debugging a 64 bit app on a 32 bit
tvd and dll on only Solaris SPARC.

r16054: Fix build failure on BigRed

r16058: Fix some typos here and there.

r16102:
Update the communicator lowest_free when we rebuild the communicator list.

r16166:
Fix a SEGV when the user updates the message queue graph after
the user executable has called MPI_Finalize(). It happens when
removing the group from each of the communicators, that MPI_COMM_NULL
doesn't have a group.  Also fix the code from skipping over every
other communicator when freeing the groups.