Changed the VERSION file to rc3, since rc2 was already posted. And:
Closes #1246: Make OS X Leopard use a more durable backtrace API

Submitted by Brian B, reviewed by Jeff, RM-Approved by tdd.

r17840:
This commit lowers the priority of the darwin backtrace component 
below that of the execinfo and stackprint components, which 
will cause OS X Leopard to use the execinfo component. execinfo 
utilizes a public API for printing the stacktrace. The darwin 
component uses some evil hacks and a not-so supported package from 
Apple to print the stack trace.