svn merge -r18930:18935 https://svn.open-mpi.org/svn/ompi/trunk/ ./
Move the latest changes related to the OB1 PML from the trunk to the 1.3.
Unfortunately, these commits contain the last part of the fix for the
64 bit segfault, and apparently I somehow mix them. Anyway, they were
supposed to go to the 1.3m but I was waiting for one more commit.

Commit r18931:
Small cleanups. Remove some switch cases that cannot be reached. Rename 
a struct field.

Commit r18932:
Add more optimizations in the case where heterogeneous support 
is not enabled.

Commit 18933:
Fix typo.

Commit r18934:
In the case where the environment is homogeneous we can ALWAYS create 
the receiver convertor when we create the request (as we know all 
architectures are identical).