Closes #1584: Fix for the C modulus problem in opal_free_list
Submitted by bosilca, Reviewed by timattox, RM-approved by bbenton.

r19778:
Solve a modulus rounding error. As the modulus can be signed (in C89 it take
the sign of the divident), we have to cast the pointer to an uintptr_t in
order to be able to correctly compute how to align it on the cache line.
Reported by Patrick Farrell and solved by Stephan Kramer. Thanks!