Closes #1663: Fix persistent sends when C/R is active
Submitted by jjhursey, Reviewed by timattox, RM-approved by bbenton

r19962:
A fix for the intel "MPI_Send_init_ator_c" test. 

It highlighted a bug in the bookmark component where for persistent
sends we were not copying the context, but just moving it. This
caused us to lose track of the message if it is started/completed
multiple times.