Closes #1742: Fix IOF problem with keyboard entry to remote procs
Submitted by rhc, Reviewed by jsquyres, RM-approved by bbenton

r20246:
Fix a reported bug whereby keyboard entry to a remote proc was
being lost after the first iteration. In other words, if an
application has a proc reading stdin from the keyboard, and
that proc is not co-located with mpirun, then the system would hang. 

The problem was eventually traced to two bugs in the code: 
 1. the orted wasn't resetting the write event flag, thus preventing
    itself from turning it on again. 
 2. the HNP needed to check if the stdin was attached to tty or
    not before adding the delay for fairness. If it is attached
    to a tty, there is no need for the delay. This prevents some
    strangely slow typing response.