Closes #1774: sm btl backing file max size setting fix
Submitted by rolfv, Reviewed by timattox, RM-approved by bbenton

r20230:
Increase the default maximum size that the sm btl file is allowed 
to grow to. Without this change, jobs with np>120 get errors. 
This does not change anything for np<16 jobs. It only comes into 
play with larger np count on a node. I imagine that this can be 
scaled back in the future if the usage of memory in the sm 
btl is improved. 

This fixes #1449.