Closes #1792: 32/64 bit Fortran handle fixes
Submitted by George / Jeff, reviewed by Jeff / George, RM-approved by bbenton

r20420:
Fix two problems with the way we handle the lvalue in the case
the Fortran and C integers have different sizes:
 1. Do not modify the read only parameter of the Fortran MPI
    interface (i.e be standard compliant).
 2. When Fortran integers are 64 bits long, don't generate unlawful code.

Thanks to Christoph van Wullen for the bug report.

r20441: Fix the compilation for Windows.

r20447: Silence a possible casting warning.

r20451: Correctly deal with the whole array.

r20462: 
{{{ s/__n/converted_n/ }}} -- according to C99, symbols that being
with {{{ "__" }}} are the domain of the compiler.