References ticket #1763: Fix -wdir option
Submitted by rhc, to be MTT-reviewed

Contains elements of r20306, r20341, and r20432.

Improve the efficiency of the launch system by changing
the outer loop to being over app_context, and adding a
flag to the app_context so the daemon can record that
"this app is on my node" when decoding the launch msg.

If the --wdir option is given, check to see if the user provided
a relative path. If so, convert it to an absolute path. This is
needed to maintain consistent behavior across environements. Some
environments automatically chdir to your current working directory
when launching the remote orted, while others (e.g., ssh) don't.
This levels the playing field and reduces user surprise.

Move the context checking code down into just prior to launching
each child app. This is necessary so that individual app context
wdir options are respected. Also, ensure that we return to our
"base" directory after each app is launched so that the relative
positions of the wdir options for each app_context are with respect
to our base directory, instead of the last wdir option.