Fixes #1821, #1969
r20706

Give a small chance to those with an "IP guru" admin-sys to define
what exactly is
a private IPv4 address. By deafult we obide to the RFC1918 and
RFC3330, but we have
the opportunity to change them.

Based on a patch from Camille Coti.

r20707

No memory leaks (this is an improvement for r20706).

r20916

Mark the address with the correct type.

r21610

Convert the port number in network order before binding the socket.
Thanks to Mariusz Mamonski (mamonski@man.poznan.pl) for the bug
report and patch.

r21611

Add a TODO bullet for later processing ...

r21612

Pretty print an error message when the specified range of ports (for
both
IPv4 and IPv6) is outside the legal boundaries. This fixes #1869.

r21616

Improve the help messages from r20706.