Fixes #2357: Move ptmalloc2 fixes to v1.4
Submitted by jsquyres, reviewed by cyeoh, RM-approved by bosilca

r22675
Fixes the problem where the rcache and core memory allocation can deadlock itself

r22863
Really fixes #2104. There is a lengthy discussion about this patch on
#2322.

The short version is that this patch consolidates two pieces of code
that call the back-end munmap and ensures that (if dlsym is used) the
corresponding dlsym is only invoked once and that the variable holding
the result is volatile.

r22867
Update copyrights.