From r22675: Fixes the problem where the rcache and core memory allocation can deadlock itself.
From r22863:
There is a lengthy discussion about this patch on #2322.
The short version is that this patch consolidates two pieces of code
that call the back-end munmap and ensures that (if dlsym is used) the
corresponding dlsym is only invoked once and that the variable holding
the result is volatile.

From r22867:
Update copyrights.

This commit closes #2358.